Reworked Packets (Packet Head & Body system)
Packets now have a constant-size 'head' and a variable-size 'body', in most cases starting with a data structure known as the 'body def'. Packet Types are now 'Body Types' which links into a table of functions allowing the packet's body data to be Marshalled (deep copied & arranged into a contiguous block of memory for sending) and Mustered (making the body def ready for reading by linking its internal pointers to the appropriate memory locations if needed).